// web/src/api/models/quota-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ApeRAG API
 * ApeRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Quota information for a specific quota type
 * @export
 * @interface QuotaInfo
 */
export interface QuotaInfo {
    /**
     * Type of quota
     * @type {string}
     * @memberof QuotaInfo
     */
    'quota_type': string;
    /**
     * Maximum allowed usage
     * @type {number}
     * @memberof QuotaInfo
     */
    'quota_limit': number;
    /**
     * Current usage count
     * @type {number}
     * @memberof QuotaInfo
     */
    'current_usage': number;
    /**
     * Remaining quota available
     * @type {number}
     * @memberof QuotaInfo
     */
    'remaining': number;
}